Update _interpreter.py
Solved variable assignment bug (when declaring/updating a value with the value of another variable, if no math operations were done the name of the variable will be taken as a string. It now works correctly).
ex. ->  store a = 3; store b = a; print(b); -> "a"
now -> store a = 3; store b = a; print(b); -> 3