Update _interpreter.py
To solve the need of a space before ";" after an array or a string, updates were done both to _parser and _lexer.
This caused another error in _interpreter. It caused getVarVal to return either a NumberNode or a Token based on the input. 
To avoid the need to check this chance in every function, a simple if statement was added in getVarVal to vary the output.